move project store to file
