+++ dat.json
@@ -1,6 +1,6 @@
 {
   "url": "dat://da9387f780f2df4bb7a67137c42c21c2566b2b74ea41a8117d6322fa904d934f/",
   "title": "Project Trackbook",
-  "description": "project trackbook",
+  "description": "A lab notebook with secure file sharing & version control!",
   "fallback_page": "/index.html"
 }
